pyproject.toml
[project]
name = "laprod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic==1.17.2",
    "authlib==1.3.2",
    "bleach>=6.3.0",
    "blinker==1.9.0",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "deprecated==1.3.1",
    "email-validator==2.2.0",
    "ffmpeg-python==0.2.0",
    "flask==3.1.2",
    "gunicorn>=23.0.0",
    "flask-limiter==4.1.1",
    "flask-login==0.6.3",
    "flask-mail==0.10.0",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "flask-talisman==1.1.0",
    "flask-wtf==1.2.2",
    "fpdf==1.7.2",
    "future==1.0.0",
    "git-filter-repo==2.47.0",
    "greenlet==3.2.4",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "librosa>=0.11.0",
    "limits==5.6.0",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "ordered-set==4.1.0",
    "packaging==25.0",
    "pandoc==2.4",
    "pillow==12.0.0",
    "plumbum==1.10.0",
    "ply==3.11",
    "psycopg2-binary>=2.9.11",
    "pydub==0.25.1",
    "python-dotenv==1.2.1",
    "python-magic",
    "python-magic-bin; sys_platform == 'win32' or sys_platform == 'darwin'",
    "rarfile>=4.2",
    "reportlab==4.4.4",
    "requests==2.32.5",
    "sqlalchemy==2.0.44",
    "stripe==13.2.0",
    "talisman>=0.1.0",
    "typing-extensions==4.15.0",
    "urllib3==2.5.0",
    "werkzeug==3.1.3",
    "wrapt==2.0.1",
    "wtforms==3.2.1",
    "apscheduler>=3.11.2",
    "scipy>=1.17.1",
    "psycopg2>=2.9.11",
    "libcst>=1.8.6",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
]
